refactor: move NewCommand/DropBenchCommand/RemoveProductionCommand logic onto Bench
Bench.drop()/remove_production() hold the teardown steps directly,
mirroring Bench.restart_processes() from a2a78cce. Bench.create_at()
delegates to a new BenchCreator (pilot/core/bench_creator.py) for the
sibling-port-discovery logic, since NewCommand runs before a Bench
object exists. All three Commands shrink to CLI wiring.

admin/backend/api/v1/benches.py's create/delete bench route handlers
now call Bench.create_at()/bench.drop() directly instead of
instantiating NewCommand/DropBenchCommand to serve an HTTP request.
